requirements.txt
numpy
paho-mqtt
opencv-python
Flask